Backend/requirements.txt
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.2.2
fastapi==0.114.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.0
psycopg2-binary==2.9.10
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
streamlit==1.40.1
streamlit-calendar==1.2.0
streamlit-modal==0.1.2
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
